Added an SSE program that adds two complex numbers.
The efficiency of this program is a bit suspect. In order to multiply the two complex numbers (4 floats) it needs to allocate space for 8 floats.
In doing this, it allows for parallel multiplication in one instruction, which might be worth it. 
The main problem is that if we have each thread doing this N times, where N is the number of columns, then we need LOADS of space for each thread. I might see if I can time this vs a more sequential alternative tomorrow.